page-visibility/visibility-on-pagehide.html

<!doctype html>
<title>Tests that visibilityState == hidden when pagehide is fired</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t) {
  var w = window.open("resources/visibility-on-pagehide.html");
  window.onmessage = t.step_func(function(event) {
    if (event.data === "loaded") {
      // The opened window finished loading. Close it to trigger pagehide.
      w.close();
      return;
    }
    // Check that the visibilityState value when pagehide was fired on the
    // opened window is "hidden".
    assert_equals(event.data, "hidden");
    t.done();
  });
});
</script>
